requirements.txt
streamlit==1.16.0
opencv-python-headless==4.5.3.56
keras==2.10.0
tensorflow==2.10.0
numpy==1.23.0
Pillow==9.0.0
